Renamed uncertainty.py back to conformal.py.
